Implement the recursion limit in the CBOR decoder API
It's very easy for a short packet to contain a very high number of
nested arrays (a packet with N bytes can be N-1 nested arrays), which
could cause stack overflow problems in devices using
cbor_value_map_find_value, so ensure that we won't crash.

The actual value needs to be chosen by the implementor depending on the
stack size of their devices and the expected usage.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>