Improve the byteswapping functions with GCC and Clang
Force the code to use the __builtin_swapXX functions if we know them to
be available. GCC added them in version 4.3 and Clang in version 3.2.
But since Apple released XCode with a 3.2 version prior to the release,
we can't check for Clang version numbers, we have to use __has_builtin.

In order to force the using of the builtins, we can't use ntohl and
ntohs, so we define our own set of macros.

This fixes the issue that glibc's ntohs is implemented with a ROR
instruction instead of BSWAP or MOVBE. It also fixes the cross-build
using Clang because glibc apparently assumes that you either have GCC or
you have x86.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>