Fixed inline function linking in C99
The CBOR_INLINE_API define previously was just 'inline'
for newer C compilers, however this was causing a failure
to link to these functions in the event that optimizations
were turned off.  This fix removes the condition on newer
compilers, and makes it ALWAYS static inline unless
the compiler is in C++ mode.

Signed-off-by: Erich Keane <erich.keane@intel.com>