API Change: make cbor_encode_{float,double} take the value by-value
Instead of by pointer, which was inconsistent with all the other
cbor_encode_xxx functions.

Now only cbor_encode_half_float is an exception, but since the C
language doesn't have a standardised half-float type, we can't have it
in the API. Even for targets where GCC has an __fp16 type, we couldn't,
as the manual says (Top > C Extensions > Half-Precision):

  In addition, you cannot declare a function with a return value or
  parameters of type `__fp16'.

But I might remove the function and force you to use
cbor_encode_floating_point.

Fixes: #3
Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>