Allow TinyCbor to continue during a OOM situation
This patch causes TinyCbor Encoder to remain in a valid-state in the situation
where it has overrun its buffer.  It does this by setting the pointer
and end variable of the encoder to NULL, making encoder->ptr - NULL (or
encoder->ptr - encoder->end) equal the amount of additional data
necessary to properly encode the package.

Signed-off-by: Erich Keane <erich.keane@intel.com>
Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>