CBOR-to-JSON: Add support for saving CBOR metadata in JSON
Metadata is saved only when the item in question is part of a map,
otherwise we don't have anywhere to save the metadata.

The metadata consists of an extra JSON object, whose key is the same key
with "$cbor" appended. The possible entries in that map are:
 - tag: the innermost tag before the actual type
 - t:   the original CBOR type (CborType)
 - v:   supplemental value to reconstruct the original value
        for simple types, it's the simple type (0-255)
        for integers, it's the sign followed by the hex dump
        for floating point, it's "nan", "inf" or "-inf"

In addition, if CborConvertStringifyMapKeys is active and the key itself
wasn't a string, an extra object entry with the same string key followed
by "$keycbordump" is added, indicating that the key was stringified.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>