Add cbor_encoder_get_buffer_size and get_extra_bytes_needed
This removes the need for people to go messing with the internals of
CborEncoder. It also provides an alternative way to determine if a
buffer OOM error occurred without checking the return value of each
function. That, of course, requires the user to be sure no other error
can happen.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>