Fix a parsing issue caused by pointer wraparounds
If the size of the string was a suitable value, the "ptr + total" check
could overflow and we wouldn't know because the sum would most likely be
smaller than value->parser->end. We would then proceed to allocate a
silly amount of memory.

I never caught this in my own testing because, on Linux, the heap
usually starts in the low 4 GB of the address space, so adding
0xffffffff000000000 would not cause an overflow. However, on OS X, the
heap starts above 4 GB, so the pointer does wrap around.

So test specifically for the pointer wraparound cases, plus the case of
sizes larger than the address space in 32-bit systems.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>