Parser: rewrite the iterate_string_chunks function with new chunk code
The new get_string_chunk function allows us to merge the chunked and
non-chunked versions into one single code path. Not only will this share
code with the chunk functions (if the user is using them), there is an
actual reduction in code size after this change.

Before, the function was 0x1d8 bytes in size on x86-64 (-march=skylake)
and 0x17b bytes for x86 (-march=silvermont -miamcu). After, the
combination of the the new iteration function and get_string_chunk is,
respectively, 0x193 and 0x157 bytes, a reduction of 69 and 36 bytes.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>