cborpretty.c: Use the new fragment API in the string dumpers
This means we don't need to allocate memory anymore, we just print
directly from the CBOR byte stream. And since the pretty dumper now uses
the fragment API, we don't need a separate test for validating it. But,
for that reason, we need to test the dup/copy API separately.

The use of the fragment API allows us to test that the dumper does
detect one UTF-8 multibyte sequence split across different string
chunks.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>