Validation: Implement recognition of known tags
This implements:
 - CborValidateNoUnknownTagsSA
 - CborValidateNoUnknownTagsSR
 - CborValidateNoUnknownTags

This commit adds a Perl script that parses a machine-readable list of
known tags (also added in this commit). That allows us to more easily
update the known tag list when more tags are added by IANA to the
registry (see API documentation for the rationale on what tags we
recognise).

If there's an interest, we can change the file where the tag list is
generated into code so that implementations can choose which tags they
want to recognise at compile time.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>