Tests: print the error message insead of just a number (or nothing)
Some QCOMPARE did convert to int() so we would see an error number, but
not all comparisons did. For those those that didn't, QCOMPARE would
just print that the values differed, but not show what.

So try to harmonize on printing the error message.

I may not have caught all uses...

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>