Pretty: limit the number of recursions, like the Parser
The pretty-dumper code did not have a limit, assuming that the input had
been validated beforehand, so if the nesting was too deep, one would
have got an error before getting to the dumper.

However, it's not inconceivable that someone adds a logger code to a
packet a bit too early. So let's be defensive and apply the limiter here
too.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>