Update docs for 0.5 (Travis build 127)
Matching commit d2603496b3ee925a3f68bf21069825f0682b0665:
Fix *some* C89/C90 issues in the code

This fixes the issues in all the headers (even in the private headers),
but not all of them in the .c files. I only fixed those that didn't make
the code look much uglier or those for which GCC complains with
-std=c90. The compiler only warns about mixed mixed variable
declarations and code, unless you use -pedantic-errors.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>