Add a make variable to pass options to compiler on configuring.
CFLAGS and LDFLAGS are not applying for tests on configuring.
(It seems to be a design for keeping a clean environment for configuring.)

But some case needs to pass options.
For example, in a case that builds with arm-gcc, mostly case needs to pass options
like as a '--specs=nosys.spec' to the compiler and linker.

Adding 'CONFIGURE_FLAGS' variable to pass options to compiler
from the make command line.
'CONFIGURE_FLAGS' also put to .config file,
So we don't need to set this variable on compiling.

Signed-off-by: TOKITA, Hiroshi <tokita.hiroshi@gmail.com>