CBOR-to-JSON: Output map-key directly to a stream
cbor_value_to_json_stream() calls cbor_value_to_pretty_stream() with the stream that given for it,
for output directly the text that representation of value.
It is no need to cache with memstream.

In map_to_json(), the map key may be written twice.
In such case, keeping iterator copy and use for re-read.

Signed-off-by: TOKITA, Hiroshi <tokita.hiroshi@gmail.com>